Improv service memory management
The idea of the mprov service is to be invoked and stopped at will
and for that all data used for the service should be allocated and
deallocated respectivelly.

For that, all the data is contained within a dinamically allocated
structure that will be allocated and pointed to by variable `data`
defined in `improv.cpp`.

To allocate the structure the method `Improv::init(name)` is used which
must be called before the `Improv::start()``mehthod. the method
`Improv::stop()` is used to free all the memory previously allocated
after stopping the bt services regarding improv.